requirements.txt
# ── Data ──────────────────────────────────────────────────────────
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
pandas-market-calendars>=4.3.0

# ── Volatility ────────────────────────────────────────────────────
arch>=6.3.0

# ── Classical ML ──────────────────────────────────────────────────
hmmlearn>=0.3.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# ── Deep Learning ─────────────────────────────────────────────────
torch>=2.0.0

# ── Quantum ML ────────────────────────────────────────────────────
pennylane>=0.36.0
pennylane-lightning>=0.36.0

# ── API ───────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ── Config / Logging ──────────────────────────────────────────────
pyyaml>=6.0
loguru>=0.7.0

# ── Testing ───────────────────────────────────────────────────────
pytest>=8.0.0
httpx>=0.27.0

# ── Plotting (outputs/figures) ────────────────────────────────────
matplotlib>=3.8.0
